# Conflicts:
#	dqoai/src/main/java/ai/dqo/connectors/jdbc/AbstractJdbcSourceConnection.java
#	dqoai/src/test/java/ai/dqo/sampledata/SampleTableMetadataForeignObjectMother.java

+++ dqoai/src/main/frontend/src/components/Connection/TableView/DataStreamListView.tsx
@@ -1,8 +1,10 @@
-import React from "react";
+import React, { useState } from "react";
 import { DataStreamBasicModel } from "../../../api";
 import Button from "../../Button";
-import SvgIcon from "../../SvgIcon";
 import { DataStreamsApi } from "../../../services/apiClient";
+import ConfirmDialog from "../../CustomTree/ConfirmDialog";
+import { Checkbox } from "@material-tailwind/react";
+import SvgIcon from "../../SvgIcon";
 
 interface IDataStreamListViewProps {
   dataStreams: DataStreamBasicModel[];
@@ -17,6 +19,9 @@ const DataStreamListView = ({
   onCreate,
   onEdit
 }: IDataStreamListViewProps) => {
+  const [open, setOpen] = useState(false);
+  const [selectedStream, setSelectedStream] = useState<DataStreamBasicModel>();
+
   const setDefaultStream = async (stream: DataStreamBasicModel) => {
     try {
       await DataStreamsApi.setDefaultDataStream(
@@ -31,7 +36,11 @@ const DataStreamListView = ({
     }
   };
 
-  const deleteStream = async (stream: DataStreamBasicModel) => {
+  const deleteStream = async (stream?: DataStreamBasicModel) => {
+    if (!stream) {
+      return;
+    }
+
     try {
       await DataStreamsApi.deleteDataStream(
         stream.connection_name || '',
@@ -43,6 +52,11 @@ const DataStreamListView = ({
     } catch (err) {
       console.error(err);
     }
+  };
+
+  const openConfirmDeleteModal = (stream: DataStreamBasicModel) => {
+    setOpen(true);
+    setSelectedStream(stream);
   };
 
   return (
@@ -59,14 +73,13 @@ const DataStreamListView = ({
         <tbody>
           {dataStreams.map((stream, index) => (
             <tr key={index}>
-              <td className="px-2 py-1 relative">
+              <td className="px-2 py-1 relative flex items-center gap-2">
                 {stream.default_data_stream && (
-                  <SvgIcon
-                    name="star"
-                    className="w-4 text-indigo-700 absolute -left-4 top-1/2 transform -translate-y-1/2"
-                  />
+                  <div className="w-5 h-5 bg-primary rounded flex items-center justify-center">
+                    <SvgIcon name="check" className="text-white" />
+                  </div>
                 )}
-                {stream.data_stream_name}
+                <span>{stream.data_stream_name}</span>
               </td>
               <td className="px-2 py-1">
                 <Button
@@ -81,7 +94,7 @@ const DataStreamListView = ({
                   label="Delete"
                   color="primary"
                   variant="text"
-                  onClick={() => deleteStream(stream)}
+                  onClick={() => openConfirmDeleteModal(stream)}
                 />
               </td>
               <td className="px-2 py-1">
@@ -103,6 +116,12 @@ const DataStreamListView = ({
         label="New data stream configuration"
         color="primary"
         onClick={onCreate}
+      />
+      <ConfirmDialog
+        open={open}
+        onClose={() => setOpen(false)}
+        message={`Are you sure to delete data stream ${selectedStream?.data_stream_name}?`}
+        onConfirm={() => deleteStream(selectedStream)}
       />
     </div>
   );
